backend/requirements.txt
# backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
alembic==1.13.1
pytest==8.2.0
httpx==0.27.0
# -- Dependências de Segurança (JWT e Hash) --
passlib[bcrypt]==1.7.4
bcrypt<4.0
python-jose[cryptography]==3.3.0
jinja2